Migrate zipmap unit test to new framework
